+++ src/bin/orga.ml
@@ -42,11 +42,9 @@ let organizations_id meth id =
   let e = Equinoxe.create ~endpoint () in
   match meth with
   | GET ->
-      if has_requiered [ id ] then
+      if has_requiered id then
         let id = Option.get id in
-        Equinoxe.Orga.get_organizations_id e ~id ()
-        |> Json.filter_error
-        |> Json.pp_r
+        Equinoxe.Orga.get_organizations_id e ~id () |> Json.pp_r
       else not_all_requiered_r [ "id" ]
   | meth -> not_supported_r meth "/organizations/{id}"
 
@@ -57,7 +55,9 @@ let organizations_t =
   let exits = default_exits in
   let man =
     man_meth
-      ~get:"Retrieve information about organizations related to the user." ()
+      ~get:
+        ("Retrieve information about organizations related to the user", [], [])
+      ()
   in
 
   Term.
@@ -68,9 +68,10 @@ let organizations_id_t =
   let doc = "Show the organization of the user referenced by the id." in
   let exits = default_exits in
   let man =
-    man_meth ~get:"Retrieve information about a specific organization" ()
+    man_meth
+      ~get:("Retrieve information about a specific organization", [ "id" ], [])
+      ()
   in
-
   let id_t =
     let doc = "The organization id" in
     Arg.(value & opt (some string) None & info [ "id" ] ~doc)